remove insomnia-config, move it inside insomnia
