--- packages/insomnia-config/.eslintignore
@@ -1,1 +1,0 @@
-dist
