--- packages/insomnia-config/package.json
@@ -1,40 +1,0 @@
-{
-  "private": true,
-  "name": "insomnia-config",
-  "version": "3.6.1-beta.4",
-  "homepage": "https://insomnia.rest",
-  "description": "Configuration for Insomnia",
-  "author": "Kong <office@konghq.com>",
-  "license": "Apache-2.0",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Kong/insomnia.git",
-    "directory": "packages/insomnia-config"
-  },
-  "bugs": {
-    "url": "https://github.com/kong/insomnia/issues"
-  },
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "bootstrap": "npm run generate",
-    "lint": "eslint . --ext .js,.ts,.tsx --cache",
-    "lint:fix": "npm run lint -- --fix",
-    "test": "jest",
-    "pregenerate": "npm run build",
-    "generate": "esr ./src/generated/generate.ts",
-    "clean": "tsc --build tsconfig.build.json --clean",
-    "prebuild": "npm run clean",
-    "build": "tsc --build tsconfig.build.json"
-  },
-  "dependencies": {
-    "@apideck/better-ajv-errors": "^0.2.6",
-    "ajv": "^8.6.2"
-  },
-  "devDependencies": {
-    "esbuild-runner": "2.2.1",
-    "jest": "^28.1.0",
-    "typescript": "^4.5.5",
-    "typescript-json-schema": "^0.50.1"
-  }
-}
