--- packages/insomnia-config/src/index.ts
@@ -1,15 +1,0 @@
-export {
-  validate,
-  ErrorResult,
-  ValidResult,
-  ValidationResult,
-  isErrorResult,
-  isValidResult,
-} from './validate';
-
-export {
-  InsomniaConfig,
-  AllowedSettings,
-  ConfigVersion,
-  INSOMNIA_CONFIG_FILENAME,
-} from './entities';
