--- packages/insomnia-config/tsconfig.build.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "composite": true,
-    "outDir": "dist",
-    "rootDir": "src",
-  },
-  "include": [
-    "**/*.d.ts",
-    "**/*.schema.json",
-    "src",
-  ],
-  "exclude": [
-    "**/*.test.ts",
-    "dist",
-  ],
-}
