--- packages/insomnia-config/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "./tsconfig.build.json",
-  "compilerOptions": {
-    "composite": false,
-    "rootDir": ".",
-  },
-  "include": [
-    "**/*.d.ts",
-    ".eslintrc.js",
-    "jest.config.js",
-    "src",
-  ],
-  "exclude": [
-    "dist",
-  ],
-}
