+++ packages/insomnia-send-request/package.json
@@ -22,7 +22,6 @@
     "hkdf": "0.0.2",
     "html-entities": "^1.3.1",
     "httpsnippet": "^1.22.0",
-    "insomnia-config": "^3.6.1-beta.4",
     "insomnia-importers": "^3.6.1-beta.4",
     "insomnia-testing": "^3.6.1-beta.4",
     "isomorphic-git": "^1.5.0",
