+++ packages/insomnia/src/utils/config/entities.ts
