+++ packages/insomnia/src/utils/config/validate.ts
